requirements.txt
flask 
torch 
transformers 
nltk 
numpy
python-docx 
pdfplumber 
python-pptx
PyPDF2
textract
